% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFvarNoVol.R
\name{BMFVAR.novol}
\alias{BMFVAR.novol}
\title{Bayesian inference of MF VAR model without SV}
\usage{
BMFVAR.novol(
  y,
  K,
  p,
  dist,
  y0 = NULL,
  idq = rep(0, K),
  freq = 3,
  prior = NULL,
  inits = NULL
)
}
\arguments{
\item{y}{The input data as matrix T x K where T is the number of observations and K is the number of variables}

\item{K}{The number of variables in BVAR model.}

\item{p}{The number of lags in BVAR model.}

\item{dist}{The variable specifies the BVAR error distribution. It should be one of
c("Gaussian","Student","Skew.Student","Skew.Student", "MT","Skew.MT","MST").}

\item{y0}{The number of observations}

\item{idq}{The indicator positions of the quarter variables}

\item{freq}{The frequency of observed quarter variables.}

\item{prior}{The prior specification of BMFVAR.}

\item{inits}{The initial values of BMFVAR.}
}
\value{
A coda object of the posterior samples.
}
\description{
Bayesian inference of MF VAR model without SV
\deqn{y_t = B x_t + SQRT(w_t) A^(-1) Sigma eps_t}
}
\examples{
\dontrun{
datagen <- sim.MFVAR.novol(dist="Gaussian", idq = c(4), p = 2, K = 4, t_max = 500)
y <- datagen$y
prior <- get_prior(y, p = 2, dist="Gaussian", SV = F)
inits <- get_init(prior)
Chain1 <- BMFVAR.novol(y, K = 5, p = 2, dist = "Gaussian", y0 = NULL, prior = prior, inits = inits)
plot(Chain1)
}
}
