% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JCvarNoVol.R
\name{JCVARF.novol}
\alias{JCVARF.novol}
\title{Bayesian inference of VAR model with factor covariance matrix}
\usage{
JCVARF.novol(y, K, p, dist, y0 = NULL, prior = NULL, inits = NULL)
}
\arguments{
\item{y}{The input data as matrix T x K where T is the number of observations and K is the number of variables}

\item{K}{The number of variables in BVAR model.}

\item{p}{The number of lags in BVAR model.}

\item{dist}{The variable specifies the BVAR error distribution. It should be one of
c("Gaussian","Student","OT").}

\item{y0}{The number of observations}

\item{prior}{The prior specification of JCVARF.}

\item{inits}{The initial values of JCVARF.}
}
\value{
A coda object of the posterior samples.
}
\description{
Bayesian inference of VAR model with factor covariance matrix
\deqn{y_t = B x_t + L f_t + SQRT(w_t) Sigma eps_t}
}
\examples{
\dontrun{
datagen <- sim.BVAR.novol(dist="Gaussian", idq = c(4), p = 2, K = 4, t_max = 500)
y <- datagen$y
prior <- get_prior(y, p = 2, dist="Gaussian", SV = F)
inits <- get_init(prior)
Chain1 <- JCVARF.novol(y, K = 5, p = 2, dist = "Gaussian", y0 = NULL, prior = prior, inits = inits)
plot(Chain1)
}
}
