% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JCvarFSV.R
\name{JCVARF.SV}
\alias{JCVARF.SV}
\title{Bayesian inference of VAR model with factor RW-SV}
\usage{
JCVARF.SV(y, K, p, dist, y0 = NULL, prior = NULL, inits = NULL)
}
\arguments{
\item{y}{The input data as matrix T x K where T is the number of observations and K is the number of variables}

\item{K}{The number of variables in JCVARF model.}

\item{p}{The number of lags in JCVARF model.}

\item{dist}{The variable specifies the JCVARF error distribution. It should be one of
c("Gaussian","Student","OT").}

\item{y0}{The number of observations}

\item{prior}{The prior specification of JCVARF.}

\item{inits}{The initial values of JCVARF.}
}
\value{
A coda object of the posterior samples.
}
\description{
Bayesian inference of VAR model with factor RW-SV
\deqn{y_t = B x_t + L f_t + SQRT(w_t) H_t eps_t}
}
\examples{
\dontrun{
datagen <- sim.VAR.SV(dist="Gaussian")
y <- datagen$y
prior <- get_prior(y, p = 2, dist="Gaussian", SV = T)
inits <- get_init(prior)
Chain1 <- JCVARF.SV(y, K = 5, p = 2, dist = "Gaussian", y0 = NULL, prior = prior, inits = inits)
plot(Chain1)
}
}
