% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFVART-package.R
\docType{package}
\name{MFVART-package}
\alias{MFVART-package}
\alias{MFVART}
\title{Bayesian VAR with Stochastic volatility, asymmetry and fat tails}
\description{
Bayesian inference of MF VAR with Stochastic volatility and fat tails
}
\details{
\tabular{ll}{
Package: \tab MFVART\cr
Type: \tab Package\cr
Version: \tab 0.0.1 \cr
Date: \tab 2024-10-01\cr
License: \tab GPL (>=2)\cr
Depends: \tab R (\eqn{\geq 3.5.0}{>= 3.5.0})\cr
Imports: \tab Rcpp \cr
Suggests: \tab  \cr
LazyLoad: \tab yes }
}
\references{
Upcoming
}
\author{
Hoang Nguyen
}
